+++ packages/react/src/component/date_picker/date_picker.stories.tsx
@@ -7,7 +7,7 @@ import { observable } from 'mobx'
 const datepickerStatus = {
   date: null,
   setDate(date: Date | null) {
-    this.date = date
+    this.date = date as any
   },
 }