+++ packages/react/src/component/time_span/time_span_picker.stories.tsx
@@ -7,27 +7,20 @@ import { observable } from 'mobx'
 const store = observable({
   date: moment().startOf('day').toDate(),
   customDate: moment().hour(10).minute(30).toDate(),
-  setDate(date) {
+  setDate(date: any) {
     this.date = date
   },
-  setCustomDate(date) {
+  setCustomDate(date: any) {
     this.customDate = date
   },
 })
 
 storiesOf('TimeSpanPicker', module)
   .add('default', () => (
-    <TimeSpanPicker
-      date={store.date}
-      onChange={(date) => store.setDate(date)}
-    />
+    <TimeSpanPicker date={store.date} onChange={(date) => store.setDate(date)} />
   ))
   .add('disabled', () => (
-    <TimeSpanPicker
-      disabled
-      date={store.date}
-      onChange={(date) => store.setDate(date)}
-    />
+    <TimeSpanPicker disabled date={store.date} onChange={(date) => store.setDate(date)} />
   ))
   .add('disabledSpan', () => (
     <TimeSpanPicker
@@ -42,24 +35,22 @@ storiesOf('TimeSpanPicker', module)
   .add('自定义时间跨度为 1h', () => (
     <TimeSpanPicker
       date={store.date}
-      max={moment().hour(23).minute(0)}
+      max={moment().hour(23).minute(0).toDate()}
       span={60 * 60 * 1000}
       onChange={(date) => store.setDate(date)}
     />
   ))
   .add('自定义 children', () => (
     <TimeSpanPicker date={store.date} onChange={(date) => store.setDate(date)}>
-      <span>
-        {store.date ? moment(store.date).format('HH:mm') : '请点击选择'}
-      </span>
+      <span>{store.date ? moment(store.date).format('HH:mm') : '请点击选择'}</span>
     </TimeSpanPicker>
   ))
 
   .add('自定义开始跟结束时间', () => (
     <TimeSpanPicker
       date={store.customDate}
-      beginTime={moment().hour(8).minute(30)}
-      endTime={moment().hour(22).minute(30)}
+      beginTime={moment().hour(8).minute(30).toDate()}
+      endTime={moment().hour(22).minute(30).toDate()}
       span={60 * 60 * 1000}
       onChange={(date) => store.setCustomDate(date)}
     />